+++ backend/canisters/user_index/impl/src/lifecycle/post_upgrade.rs
@@ -1,10 +1,11 @@
 use crate::lifecycle::{init_state, reseed_rng, UPGRADE_BUFFER_SIZE};
 use crate::memory::get_upgrades_memory;
-use crate::Data;
+use crate::{Data, ONE_GB, ONE_MB};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk_macros::post_upgrade;
 use ic_stable_structures::reader::{BufferedReader, Reader};
+use open_storage_index_canister::add_or_update_users::UserConfig;
 use std::time::Duration;
 use tracing::info;
 use user_index_canister::post_upgrade::Args;
@@ -22,9 +23,18 @@ fn post_upgrade(args: Args) {
 
     let (mut data, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>) = serializer::deserialize(reader).unwrap();
 
-    canister_logger::init_with_logs(data.test_mode, logs, traces);
+    // Push all users to OpenStorage
+    // Diamond members get 1GB, all others get 100MB
+    let now = env.now();
+    for user in data.users.iter() {
+        let storage_limit = if user.diamond_membership_details.is_active(now) { ONE_GB } else { 100 * ONE_MB };
+        data.open_storage_user_sync_queue.push(UserConfig {
+            user_id: user.principal,
+            byte_limit: storage_limit,
+        })
+    }
 
-    data.users.initialize_diamond_members(env.now());
+    canister_logger::init_with_logs(data.test_mode, logs, traces);
 
     init_cycles_dispenser_client(data.cycles_dispenser_canister_id);
     init_state(env, data, args.wasm_version);